fix config and reads

+++ run.sh
@@ -26,18 +26,29 @@ LIB_PATH=../Quark/quarkstore/quarklibio/build
 
 ENGINE=fs
 DIR=./dbdir
+DISTRIBUTION=zipfian
+PREPOPULATE=65536
+WRITE_SIZE=0
+READ_SIZE=100000
+KEY_SIZE=12
 
 if [[ "$DIR" == "quarkstore" ]]; then
     InstallQuark
 else
     mkdir -p $DIR
-    rm -rf $DIR/*
+    if [ "$PREPOPULATE" -gt 0 ]; then
+        rm -rf $DIR/*
+    fi
 fi
 
 LD_LIBRARY_PATH=$LIB_PATH ./main \
     --engine=$ENGINE \
     --db-directory=$DIR \
-    --prepopulate-size=65536
+    --workload-distribution=$DISTRIBUTION \
+    --prepopulate-size=$PREPOPULATE \
+    --write-size=$WRITE_SIZE \
+    --read-size=$READ_SIZE \
+    --key-size=$KEY_SIZE
 
 ret=0
